fix ctidy

+++ tasks/kurpiakov_a_elem_vec_sum/common/include/common.hpp
@@ -3,12 +3,13 @@
 #include <string>
 #include <tuple>
 #include <vector>
+#include <cstdint>
 
 #include "task/include/task.hpp"
 
 namespace kurpiakov_a_elem_vec_sum {
 using InType = std::tuple<int, std::vector<int>>;
-using OutType = long long;
+using OutType = int64_t;
 using TestType = std::tuple<InType, std::string, OutType>;
 using BaseTask = ppc::task::Task<InType, OutType>;
